fix integer size conversion warnings
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@18768 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseMerge/LocatorBar.cpp
@@ -272,7 +272,7 @@ void CLocatorBar::PaintView(CDC& cacheDC, CBaseView* view, CDWordArray& indents,
 		{
 			if (view->m_arMarkedWordLines[i])
 			{
-				cacheDC.FillSolidRect(rect.left + (width*stripeIndex/3), height*i/m_nLines, 
+				cacheDC.FillSolidRect(rect.left + (width*stripeIndex/3), (int)(height*i/m_nLines), 
 					barwidth, max(height/m_nLines,2), color);
 			}
 		}
